init websocket server
