+++ src/actions.c
@@ -1,0 +1,13 @@
+#include "../include/actions.h" 
+
+void sendCocktails(ws_cli_conn_t *client){
+    printf("nothing\n");
+}
+
+void sendBottles(ws_cli_conn_t *client){
+    printf("nothing\n");
+}
+
+void processOrder(ws_cli_conn_t *client, cJSON *json_order){
+    printf("nothing\n");
+}
